08_semestre/pdytr/practica/tp_3/ejer4/requirements.txt
grpc-stubs==1.53.0.5
grpcio==1.66.2
grpcio-tools==1.66.2
numpy==2.1.2
pandas==2.2.3
protobuf==5.28.2
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==75.1.0
six==1.16.0
tzdata==2024.2
